+++ menu/build.gradle
@@ -2,15 +2,10 @@ plugins {
     id 'org.springframework.boot' version '2.3.3.RELEASE'
     id 'io.spring.dependency-management' version '1.0.10.RELEASE'
     id 'java'
-    id 'application'
 }
 
-application {
-    mainClass = 'kitchenpos.Application'
-}
-
-group 'camp.nextstep.edu'
-version '0.0.1-SNAPSHOT'
+group = 'camp.nextstep.edu'
+version = '1.0'
 
 sourceCompatibility = JavaVersion.VERSION_11
 targetCompatibility = JavaVersion.VERSION_11
@@ -20,13 +15,11 @@ repositories {
 }
 
 dependencies {
-    implementation project(':table')
-    implementation project(':menu')
-
+    implementation project(':product')
+    
     implementation 'org.springframework.boot:spring-boot-starter-actuator'
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
     implementation 'org.springframework.boot:spring-boot-starter-web'
-    implementation 'org.flywaydb:flyway-core'
     implementation 'org.springframework.boot:spring-boot-starter-validation'
     runtimeOnly 'com.h2database:h2'
     testImplementation('org.springframework.boot:spring-boot-starter-test') {
