+++ menu/src/main/java/kitchenpos/menu/ui/MenuRestController.java
@@ -2,9 +2,9 @@ package kitchenpos.menu.ui;
 
 import java.net.URI;
 import kitchenpos.menu.application.MenuService;
-import kitchenpos.menu.application.dto.MenuRequest;
 import kitchenpos.menu.application.dto.MenuResponse;
 import kitchenpos.menu.application.dto.MenuResponses;
+import kitchenpos.menu.application.dto.MenuRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
