+++ menu/src/main/java/kitchenpos/menugroup/application/MenuGroupService.java
@@ -1,8 +1,8 @@
 package kitchenpos.menugroup.application;
 
 import kitchenpos.menugroup.application.dto.MenuGroupRequest;
-import kitchenpos.menugroup.application.dto.MenuGroupResponse;
 import kitchenpos.menugroup.application.dto.MenuGroupResponses;
+import kitchenpos.menugroup.application.dto.MenuGroupResponse;
 import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.menugroup.domain.repository.MenuGroupRepository;
 import org.springframework.stereotype.Service;
