+++ menu/src/test/java/kitchenpos/ApplicationTest.java
@@ -1,0 +1,9 @@
+package kitchenpos;
+
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ApplicationTest {
+    public void contextLoads() {}
+}
+
