+++ menu/src/test/java/kitchenpos/menu/ui/MenuRestControllerTest.java
@@ -1,5 +1,4 @@
-package kitchenpos.menugroup.ui;
-
+package kitchenpos.menu.ui;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
@@ -10,12 +9,12 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.Collections;
-import kitchenpos.TestFixtures;
-import kitchenpos.menugroup.application.MenuGroupService;
-import kitchenpos.menugroup.application.dto.MenuGroupRequest;
-import kitchenpos.menugroup.application.dto.MenuGroupResponse;
-import kitchenpos.menugroup.application.dto.MenuGroupResponses;
-import kitchenpos.menugroup.domain.MenuGroup;
+import kitchenpos.MenuFixture;
+import kitchenpos.menu.application.MenuService;
+import kitchenpos.menu.application.dto.MenuResponse;
+import kitchenpos.menu.application.dto.MenuResponses;
+import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.ui.MenuRestController;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -26,9 +25,9 @@ import org.springframework.http.MediaType;
 import org.springframework.mock.web.MockHttpServletResponse;
 import org.springframework.test.web.servlet.MockMvc;
 
-@DisplayName("메뉴 그룹 api")
-@WebMvcTest(MenuGroupRestController.class)
-class MenuGroupRestControllerTest {
+@DisplayName("메뉴 api")
+@WebMvcTest(MenuRestController.class)
+class MenuRestControllerTest {
     @Autowired
     private MockMvc mockMvc;
 
@@ -36,15 +35,16 @@ class MenuGroupRestControllerTest {
     private ObjectMapper objectMapper;
 
     @MockBean
-    private MenuGroupService menuGroupService;
+    private MenuService menuService;
 
     @Test
     void create() throws Exception {
-        final String content = objectMapper.writeValueAsString(new MenuGroupRequest("메뉴그룹이름"));
-        final MenuGroup menuGroup = TestFixtures.createMenuGroup();
-        when(menuGroupService.create(any())).thenReturn(new MenuGroupResponse(menuGroup));
+        final Menu menu = MenuFixture.createMenu();
+        menu.updateMenuProducts(Collections.singletonList(MenuFixture.createMenuProduct()));
+        final String content = objectMapper.writeValueAsString(MenuFixture.createMenuRequest(menu));
+        when(menuService.create(any())).thenReturn(new MenuResponse(menu));
 
-        final MockHttpServletResponse response = mockMvc.perform(post("/api/menu-groups")
+        final MockHttpServletResponse response = mockMvc.perform(post("/api/menus")
                         .content(content)
                         .contentType(MediaType.APPLICATION_JSON))
                 .andReturn()
@@ -52,17 +52,17 @@ class MenuGroupRestControllerTest {
 
         assertAll(
                 () -> assertThat(response.getStatus()).isEqualTo(HttpStatus.CREATED.value()),
-                () -> assertThat(response.getHeader("Location")).isEqualTo("/api/menu-groups/" + menuGroup.getId())
+                () -> assertThat(response.getHeader("Location")).isEqualTo("/api/menus/" + menu.getId())
         );
     }
 
     @Test
     void list() throws Exception {
-        when(menuGroupService.list()).thenReturn(
-                new MenuGroupResponses(Collections.singletonList(TestFixtures.createMenuGroup()))
-        );
+        final Menu menu = MenuFixture.createMenu();
+        menu.updateMenuProducts(Collections.singletonList(MenuFixture.createMenuProduct()));
+        when(menuService.list()).thenReturn(new MenuResponses(Collections.singletonList(menu)));
 
-        final MockHttpServletResponse response = mockMvc.perform(get("/api/menu-groups"))
+        final MockHttpServletResponse response = mockMvc.perform(get("/api/menus"))
                 .andReturn()
                 .getResponse();
