+++ menu/src/test/java/kitchenpos/menugroup/application/MenuGroupServiceTest.java
@@ -7,6 +7,7 @@ import static org.mockito.Mockito.when;
 
 import java.util.Arrays;
 import java.util.List;
+import kitchenpos.menugroup.application.MenuGroupService;
 import kitchenpos.menugroup.application.dto.MenuGroupRequest;
 import kitchenpos.menugroup.application.dto.MenuGroupResponse;
 import kitchenpos.menugroup.application.dto.MenuGroupResponses;
