+++ settings.gradle
@@ -2,4 +2,5 @@ rootProject.name = 'kitchenpos'
 include 'app'
 include 'product'
 include 'table'
+include 'menu'
